Improvements with automatic data loading
No longer need to run 'dotnet ef database update' to create database and execute the initial data script manually.

Now a check is being made if the database exists in the first request to the database. 

If case does not exist, it will create the database with the tables automatically. Also it will read files from the scripts folder to create procedures and then run them to insert the initial data.